Split the catalog and descriptor IO out of the loader
Give each artifact its own reader and writer module, catalog_io and
descriptor_io, following the split the seabed package already uses. The
loader is left holding the two readers for the older deployment config
files, which are on their way out.

Trim the moved docstrings to their summary and sections; the rationale
they carried belongs to the design discussion rather than the module.
