+++ src/afft/deployment/__init__.py
@@ -10,6 +10,10 @@ from .catalog import (
     CatalogVesselProfile as CatalogVesselProfile,
     DeploymentCatalog as DeploymentCatalog,
 )
+from .catalog_io import (
+    read_deployment_catalog as read_deployment_catalog,
+    write_deployment_catalog as write_deployment_catalog,
+)
 from .descriptor import (
     DeploymentDescriptor as DeploymentDescriptor,
     DeploymentFileSection as DeploymentFileSection,
@@ -24,17 +28,17 @@ from .descriptor import (
     VesselIdentity as VesselIdentity,
     VesselSensor as VesselSensor,
 )
+from .descriptor_io import (
+    read_deployment_descriptors as read_deployment_descriptors,
+    write_deployment_descriptors as write_deployment_descriptors,
+)
 from .files import (
     DeploymentFiles as DeploymentFiles,
     collect_deployment_files as collect_deployment_files,
 )
 from .loader import (
     load_deployment_config as load_deployment_config,
-    read_deployment_catalog as read_deployment_catalog,
-    read_deployment_descriptors as read_deployment_descriptors,
     read_deployment_info as read_deployment_info,
-    write_deployment_catalog as write_deployment_catalog,
-    write_deployment_descriptors as write_deployment_descriptors,
 )
 from .types import (
     DeploymentConfig as DeploymentConfig,
